+++ tests/Transport/MimeTypeResolver/CustomMimeTypeResolverTest.php
@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 namespace Phptg\BotApi\Tests\Transport\MimeTypeResolver;
 
-use Phptg\BotApi\Tests\Support\StubResourceReader;
-use Phptg\BotApi\Transport\InputFileData;
-use Phptg\BotApi\Transport\ResourceReader\NativeResourceReader;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Phptg\BotApi\Transport\MimeTypeResolver\CustomMimeTypeResolver;
@@ -18,11 +15,11 @@ final class CustomMimeTypeResolverTest extends TestCase
 {
     public static function dataBase(): iterable
     {
-        yield [null, new InputFile(null)];
-        yield [null, InputFile::fromLocalFile(__DIR__ . '/files/test.unknown')];
-        yield ['text/plain', InputFile::fromLocalFile(__DIR__ . '/files/test.txt')];
-        yield ['text/css', InputFile::fromLocalFile(__DIR__ . '/files/test.txt', 'test.css')];
-        yield ['text/css', InputFile::fromLocalFile(__DIR__ . '/files/test.txt', 'test.CSS')];
+        yield [null, new InputFile(fopen('php://memory', 'rb'))];
+        yield [null, new InputFile(__DIR__ . '/files/test.unknown')];
+        yield ['text/plain', new InputFile(__DIR__ . '/files/test.txt')];
+        yield ['text/css', new InputFile(__DIR__ . '/files/test.txt', 'test.css')];
+        yield ['text/css', new InputFile(__DIR__ . '/files/test.txt', 'test.CSS')];
     }
 
     #[DataProvider('dataBase')]
@@ -33,16 +30,7 @@ final class CustomMimeTypeResolverTest extends TestCase
             'txt' => 'text/plain',
             'css' => 'text/css',
         ]);
-        $fileData = new InputFileData(
-            $file,
-            [
-                new NativeResourceReader(),
-                new StubResourceReader(uri: 'custom://test'),
-            ],
-        );
 
-        $result = $resolver->resolve($fileData);
-
-        assertSame($expected, $result);
+        assertSame($expected, $resolver->resolve($file));
     }
 }